fix: attempt to improve the quality of line highlighting matches for invalid variables
eg:
```
{#      <title>{% block title %}{{ name }}...{% endblock %}</title>#}
      <title>{% block title %}{{ name }}Django REST framework{% endblock %}</title>
```
+ a child template including:
```
{{ form.username }}
```
the first line of the parent template should not be highlighted, because it's in a line comment.
the line from the child template should also not be highlighted, because it's not the correct match...

It's all still going to be subject to false positives, I'm sure. But hey ho.